p2pv2: buffer and merge signaling chunks at the bridge level
In P2Pv2, package numbers are not random numbers and always begin with 0. So there are high chances that different bridges receive the same pacakge number (blob id). It's better to buffer signaling messages directly at bridge level.
PS: We assume that we always receive a complete SLP message on a single bridge