Implement pending contact signal
Currently there is no way to get notified if we got invited by
some other user on the network. This patch implements a new signal,
contact-pending, that might be used to get notified of pending
contacts. This is really useful for things like automatic bots
which should auto-accept this contacts, or IM desktop clients,
which popup a dialog when somebody invites us.

https://bugs.freedesktop.org/show_bug.cgi?id=29132