sb: don't set the switchboard in the handler constructor
We don't request the switchboard in the handler constructor anymore (too much overhead). That means you can go back to set the switchboard right after creating the handler, no need to pass the switchboard as a constructor argument.

The reason it was that way was to avoid requesting a new switchboard when we were handling a message.