p2p: buffer data and SLP blobs in transport manager
The key to find which blob to append the chunk to has been changed to (peer, peer_guid, session_id). This way, we should be able to identify the right blob no matter if it's a SLP blob or data one. The key is now used as well to register data buffer and to blacklist session. The same session ID can now be used by more than one peer.